requirements.txt
requests
python-dotenv
pandas
matplotlib
PyMuPDF==1.22.0
pdfplumber